requirements.txt
fastapi==0.101.1
uvicorn[standard]==0.22.0
python-multipart==0.0.6
langchain==0.0.400
openai==1.8.5
faiss-cpu==1.7.4
tiktoken==0.4.0
python-dotenv==1.0.0
unstructured==0.7.9
python-docx==0.8.11
aiofiles==23.1.0
requests==2.31.0